effective/requirements.txt
asgiref==3.10.0
bcrypt==5.0.0
Django==5.2.7
djangorestframework==3.16.1
PyJWT==2.10.1
sqlparse==0.5.3
